GlobalProfile moved to 'global.profile' package
