+++ src/main/java/egl/client/model/global/profile/GlobalProfile.java
@@ -1,4 +1,4 @@
-package egl.client.model.profile;
+package egl.client.model.global.profile;
 
 import egl.core.model.profile.Profile;
 import lombok.Data;
