+++ src/main/java/egl/client/repository/profile/GlobalCredentialsRepository.java
@@ -1,6 +1,6 @@
 package egl.client.repository.profile;
 
-import egl.client.model.profile.GlobalCredentials;
+import egl.client.model.global.profile.GlobalCredentials;
 import egl.client.repository.EntityRepository;
 import egl.core.model.profile.Profile;
 import org.springframework.stereotype.Repository;
