+++ src/main/java/egl/client/repository/profile/GlobalProfileRepository.java
@@ -1,6 +1,6 @@
 package egl.client.repository.profile;
 
-import egl.client.model.profile.GlobalProfile;
+import egl.client.model.global.profile.GlobalProfile;
 import egl.client.repository.EntityRepository;
 import org.springframework.stereotype.Repository;
