+++ src/main/java/egl/client/service/model/profile/GlobalProfileService.java
@@ -1,6 +1,6 @@
 package egl.client.service.model.profile;
 
-import egl.client.model.profile.GlobalProfile;
+import egl.client.model.global.profile.GlobalProfile;
 import egl.client.repository.profile.GlobalProfileRepository;
 import org.springframework.stereotype.Service;
